backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
bcrypt==4.2.1
PyJWT==2.10.1
cryptography==45.0.7
Pillow==11.1.0
pytest==8.3.4
httpx==0.28.1
